% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvblocker.R
\name{check_cvblocks}
\alias{check_cvblocks}
\title{Build descriptive table of CV blocks}
\usage{
check_cvblocks(x, blocks)
}
\arguments{
\item{x}{A vector of labels.}

\item{blocks}{A vector of cv blocks.}
}
\value{
A matrix with a row for each cv block and a column for each condition label, along with margin totals and grand total.
}
\description{
Build descriptive table of CV blocks
}
\examples{
x <- rep(c(0,1,2), c(49,949,874))
x <- x[sample(length(x))]
blocks <- cvblocker(x, k=10)
M <- check_cvblocks(x, blocks)
print(M)
}
