% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvblocker.R
\name{cvblocker}
\alias{cvblocker}
\title{Split labels into k blocks for cross validation}
\usage{
cvblocker(x, k)
}
\arguments{
\item{x}{A vector of labels.}

\item{k}{Number of blocks to compose.}
}
\value{
A vector of block labels 1:\code{k}, of for each element of \code{x}.
}
\description{
Split labels into k blocks for cross validation
}
\examples{
x <- rep(c(0,1,2), c(49,949,874))
x <- x[sample(length(x))]
blocks <- cvblocker(x, k=10)
M <- check_cvblocks(x, blocks)
print(M)
}
